[BUGFIX] Fix NS not being sent on the home link
Commit 454c91bf787 has introduced an issue: MN could not send NS
from the Home Link anymore. Using a MAIN policy for the NS/NA bypass
policy would make the SUB block policy (upon BU/BA exchange) to be
enforced. This commit fixes the issue by installing both a MAIN and
SUB bypass policy of higher preference.

Using a MAIN block policy instead of the SUB one is not a solution
since it would block the sending of BUs.

Signed-off-by: Romain Kuntz <r.kuntz@ipflavors.com>